+++ apps/prairielearn/src/lib/editors.sql
@@ -55,3 +55,17 @@ WHERE
   id = $course_id
 RETURNING
   draft_number;
+
+-- BLOCK select_assessments_with_assessment_set
+SELECT
+  ci.short_name AS course_instance_directory,
+  a.tid AS assessment_directory
+FROM
+  assessments AS a
+  JOIN assessment_sets AS aset ON (aset.id = a.assessment_set_id)
+  JOIN course_instances AS ci ON (ci.id = a.course_instance_id)
+WHERE
+  aset.name = $assessment_set_name
+  AND aset.course_id = $course_id
+  AND a.deleted_at IS NULL
+  AND ci.deleted_at IS NULL;