+++ apps/prairielearn/src/pages/instructorAssessmentQuestions/instructorAssessmentQuestions.tsx
@@ -1,6 +1,5 @@
 import * as path from 'path';
 
-import sha256 from 'crypto-js/sha256.js';
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
 import fs from 'fs-extra';
@@ -21,7 +20,7 @@ import {
   StaffTagSchema,
   StaffTopicSchema,
 } from '../../lib/client/safe-db-types.js';
-import { FileModifyEditor } from '../../lib/editors.js';
+import { FileModifyEditor, getOriginalHash } from '../../lib/editors.js';
 import { features } from '../../lib/features/index.js';
 import { getPaths } from '../../lib/instructorFiles.js';
 import { formatJsonWithPrettier } from '../../lib/prettier.js';
@@ -30,8 +29,9 @@ import { resetVariantsForAssessmentQuestion } from '../../models/variant.js';
 import { ZoneAssessmentJsonSchema } from '../../schemas/infoAssessment.js';
 
 import { InstructorAssessmentQuestionsTable } from './components/InstructorAssessmentQuestionsTable.js';
+import { InstructorAssessmentQuestionsTableLegacy } from './components/InstructorAssessmentQuestionsTableLegacy.js';
 import type { CourseQuestionForPicker } from './types.js';
-import { stripZoneDefaults } from './utils/dataTransform.js';
+import { serializeZonesForJson } from './utils/dataTransform.js';
 import { buildHierarchicalAssessment } from './utils/questions.js';
 
 const router = Router();
@@ -57,6 +57,10 @@ const SaveQuestionsSchema = z.object({
 router.get(
   '/',
   asyncHandler(async (req, res) => {
+    if (!res.locals.authz_data.has_course_permission_preview) {
+      throw new HttpStatusError(403, 'Access denied (must be course previewer)');
+    }
+
     const [questionRows, courseQuestions] = await Promise.all([
       selectAssessmentQuestions({
         assessment_id: res.locals.assessment.id,
@@ -73,14 +77,7 @@ router.get(
       'infoAssessment.json',
     );
 
-    const assessmentPathExists = await fs.pathExists(assessmentPath);
-
-    let origHash = '';
-    if (assessmentPathExists) {
-      // TODO: Use helper once assessment sets PR lands
-      const assessmentFileContents = await fs.readFile(assessmentPath, 'utf8');
-      origHash = sha256(b64EncodeUnicode(assessmentFileContents)).toString();
-    }
+    const origHash = (await getOriginalHash(assessmentPath)) ?? '';
 
     // We use the database instead of the contents on disk as we want to consider the database as the 'source of truth'
     // for doing operations.
@@ -123,20 +120,35 @@ router.get(
         },
         content: (
           <Hydrate>
-            <InstructorAssessmentQuestionsTable
-              course={pageContext.course}
-              questionRows={questionRows}
-              courseQuestions={courseQuestionsForPicker}
-              jsonZones={jsonZones}
-              urlPrefix={pageContext.urlPrefix}
-              assessment={pageContext.assessment}
-              assessmentSetName={pageContext.assessment_set.name}
-              hasCoursePermissionPreview={pageContext.authz_data.has_course_permission_preview}
-              canEdit={canEdit ?? false}
-              csrfToken={res.locals.__csrf_token}
-              origHash={origHash}
-              editorEnabled={editorEnabled}
-            />
+            {editorEnabled ? (
+              <InstructorAssessmentQuestionsTable
+                course={pageContext.course}
+                questionRows={questionRows}
+                courseQuestions={courseQuestionsForPicker}
+                jsonZones={jsonZones}
+                urlPrefix={pageContext.urlPrefix}
+                assessment={pageContext.assessment}
+                assessmentSetName={pageContext.assessment_set.name}
+                hasCoursePermissionPreview={pageContext.authz_data.has_course_permission_preview}
+                canEdit={canEdit ?? false}
+                csrfToken={res.locals.__csrf_token}
+                origHash={origHash}
+              />
+            ) : (
+              <InstructorAssessmentQuestionsTableLegacy
+                course={pageContext.course}
+                questionRows={questionRows}
+                urlPrefix={pageContext.urlPrefix}
+                assessmentType={pageContext.assessment.type}
+                assessmentSetName={pageContext.assessment_set.name}
+                assessmentNumber={pageContext.assessment.number}
+                hasCoursePermissionPreview={pageContext.authz_data.has_course_permission_preview}
+                hasCourseInstancePermissionEdit={
+                  pageContext.authz_data.has_course_instance_permission_edit ?? false
+                }
+                csrfToken={res.locals.__csrf_token}
+              />
+            )}
           </Hydrate>
         ),
       }),
@@ -147,9 +159,8 @@ router.get(
 router.get(
   '/question.json',
   asyncHandler(async (req, res) => {
-    // TODO: Is this needed?
     if (!res.locals.authz_data.has_course_permission_preview) {
-      throw new HttpStatusError(403, 'Access denied');
+      throw new HttpStatusError(403, 'Access denied (must be course previewer)');
     }
 
     const parsedQuery = z
@@ -179,6 +190,8 @@ router.post(
   '/',
   asyncHandler(async (req, res) => {
     if (req.body.__action === 'reset_question_variants') {
+      // TODO: What permission do we need to reset question variants?
+
       if (res.locals.assessment.type === 'Exam') {
         // See https://github.com/PrairieLearn/PrairieLearn/issues/12977
         throw new HttpStatusError(403, 'Cannot reset variants for Exam assessments');
@@ -197,6 +210,10 @@ router.post(
       );
       if (!editorEnabled) {
         throw new HttpStatusError(403, 'Assessment questions editor feature is not enabled');
+      }
+
+      if (!res.locals.authz_data.has_course_permission_edit) {
+        throw new HttpStatusError(403, 'Access denied (must be course editor)');
       }
 
       const body = SaveQuestionsSchema.parse(req.body);
@@ -218,7 +235,7 @@ router.post(
       const assessmentInfo = JSON.parse(await fs.readFile(assessmentPath, 'utf8'));
 
       // Strip default values from zones data.
-      const filteredZones = stripZoneDefaults(body.zones);
+      const filteredZones = serializeZonesForJson(body.zones);
 
       // Update the zones with the filtered data
       assessmentInfo.zones = filteredZones;