+++ packages/flash/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prairielearn/flash
 
+## 3.0.1
+
+### Patch Changes
+
+- 8bdf6ea: Upgrade all JavaScript dependencies
+- Updated dependencies [8bdf6ea]
+  - @prairielearn/html@5.0.1
+
 ## 3.0.0
 
 ### Major Changes