+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @prairielearn/migrations
 
+## 5.0.1
+
+### Patch Changes
+
+- 8bdf6ea: Upgrade all JavaScript dependencies
+- Updated dependencies [8bdf6ea]
+  - @prairielearn/named-locks@4.0.1
+  - @prairielearn/postgres@5.0.2
+  - @prairielearn/logger@3.0.1
+  - @prairielearn/error@3.0.2
+
 ## 5.0.0
 
 ### Major Changes